chore: refactor BasketOverview
