+++ web-registry/src/components/organisms/BasketOverview/BasketOverview.GridItem.tsx
@@ -1,0 +1,24 @@
+import { Box, Grid } from '@mui/material';
+
+import { Label } from 'web-components/lib/components/typography';
+
+interface GridItemProps {
+  label: string;
+  children: React.ReactNode;
+}
+
+export const BasketGridItem = ({
+  label,
+  children,
+}: GridItemProps): JSX.Element => {
+  return (
+    <Grid item xs={12} sm={6}>
+      <Box sx={{ mt: 4 }}>
+        <Label size="xs" sx={{ color: 'primary.contrastText', mb: 2 }}>
+          {label}
+        </Label>
+        {children}
+      </Box>
+    </Grid>
+  );
+};
