+++ web-registry/src/components/organisms/BasketOverview/BasketOverview.types.ts
@@ -1,0 +1,9 @@
+export type CreditClass = {
+  id: string;
+  name: string;
+};
+
+export type Curator = {
+  name: string;
+  address: string;
+};
